BUGFIX: Extract shared `ProjectionIntegrityViolationDetectionTrait` from Dbal specific implemetation
- removes all autoloading hacks
- allows custom content graph implementations to be tested without DBAL adapter and autoloading issues
- (from last commit) objects.yaml determines which  ProjectionIntegrityViolationDetection to use
