+++ Neos.ContentGraph.DoctrineDbalAdapter/Tests/Behavior/Features/Bootstrap/FeatureContext.php
@@ -11,12 +11,9 @@ declare(strict_types=1);
  * source code.
  */
 
-require_once(__DIR__ . '/ProjectionIntegrityViolationDetectionTrait.php');
-
 use Behat\Behat\Context\Context as BehatContext;
 use Behat\Behat\Hook\Scope\BeforeScenarioScope;
 use Neos\Behat\FlowBootstrapTrait;
-use Neos\ContentGraph\DoctrineDbalAdapter\Tests\Behavior\Features\Bootstrap\ProjectionIntegrityViolationDetectionTrait;
 use Neos\ContentRepository\Core\ContentRepository;
 use Neos\ContentRepository\Core\Factory\ContentRepositoryServiceFactoryInterface;
 use Neos\ContentRepository\Core\Factory\ContentRepositoryServiceInterface;
@@ -33,7 +30,7 @@ use Neos\ContentRepositoryRegistry\ContentRepositoryRegistry;
 class FeatureContext implements BehatContext
 {
     use FlowBootstrapTrait;
-    use ProjectionIntegrityViolationDetectionTrait;
+    use DoctrineDbalProjectionIntegrityViolatorTrait;
     use CRTestSuiteTrait;
     use CRBehavioralTestsSubjectProvider;
